% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{china_line_data}
\alias{china_line_data}
\title{a data.frame which is the border map data of China from http://nfgis.nsdi.gov.cn}
\format{a data.frame}
\usage{
china_line_data
}
\description{
a data.frame which is the border map data of China from http://nfgis.nsdi.gov.cn
}
\keyword{datasets}
