% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDetailedNumber.R
\name{get_2019nCoV}
\alias{get_2019nCoV}
\title{Getting R2019-nCoV object and checking real-time monitoring numbers of the 2019-nCoV virus}
\usage{
get_2019nCoV()
}
\value{
a R2019-nCoV object for further use
}
\description{
Getting R2019-nCoV object and checking real-time monitoring numbers of the 2019-nCoV virus
}
\examples{

x <- get_2019nCoV()
}
