% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDetailedNumber.R
\name{getDataFrame}
\alias{getDataFrame}
\title{Getting data.frame of detailed real-time statistical numbers of every provinces of China of 2019-nCoV virus cases}
\usage{
getDataFrame(x)
}
\arguments{
\item{x}{a R2019-nCoV object, i.e. the result of function get_2019nCoV}
}
\value{
a data.frame object which contains detailed real-time statistics of 2019-nCoV virus cases
}
\description{
Getting data.frame of detailed real-time statistical numbers of every provinces of China of 2019-nCoV virus cases
}
\examples{

x <- get_2019nCoV()
df <- getDataFrame(x)
}
