% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMap.R
\name{getMap}
\alias{getMap}
\title{Getting real-time monitoring map of the 2019-nCoV virus}
\usage{
getMap(x)
}
\arguments{
\item{x}{a R2019-nCoV object, i.e. the result of function get_2019nCoV}
}
\value{
a ggplot2 object which shows the real-time monitoring map of the 2019-nCoV virus
}
\description{
Getting real-time monitoring map of the 2019-nCoV virus
}
\examples{

x <- get_2019nCoV()
map <- getMap(x)
}
