% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeq.R
\name{getSeq}
\alias{getSeq}
\title{Getting the gene sequence of 2019-nCoV virus (i.e.GenBank:MN908947)}
\usage{
getSeq()
}
\value{
a string which is the gene sequence of 2019-nCoV
}
\description{
Getting the gene sequence of 2019-nCoV virus (i.e.GenBank:MN908947)
}
\examples{

Seq <- getSeq()
}
